requirements.txt
baresipy>=0.1.4
click>=8.1.3
click_loglevel
hashring>=1.5.1
prettytable>=3.3.0
pydantic>=1.9.1
pyfestival>=0.6
toml>=0.10.2
twitchAPI>=2.5.7.1
